Quiet a variety of warnings in Xcode mostly having to do with NSUInteger/NSInteger types in format strings, but also including some Xcode version checks in 4.5 such as advising to update the default debugger.

+++ IMBFlickrNode.h
@@ -156,7 +156,7 @@ typedef enum {
 
 - (NSDictionary*) argumentsForFlickrCall;
 + (NSString *)base58EncodedValue:(long long)num;
-+ (NSString *)descriptionOfLicense:(int)aLicenseNumber;
++ (NSString *)descriptionOfLicense:(NSInteger)aLicenseNumber;
 + (NSString*) identifierWithQueryParams: (NSDictionary*) inQueryParams;
 - (void) readPropertiesFromDictionary: (NSDictionary*) dictionary;
